fix: add hash and modules logic
