+++ src/helpers/getPath.js
@@ -1,10 +1,10 @@
 import path from "path";
 import { fileURLToPath } from "url";
 
-export const getDirname = (url) => {
+export const getPath = (url) => {
   const __filename = fileURLToPath(url);
   const __dirname = path.dirname(__filename);
   console.log(__dirname);
 
-  return __dirname;
+  return { __dirname, __filename };
 };
